refactor: Quite a few improvements/rewrites. (#46)
* refactor: Rewrite proof for `direct_sum_unitary`.

* fix(WFHelpers): Revert `type of` change.

Explicitly matching the types like this seems to yield better results.

* refactor: Use new `control_decomp`.

* feat: Prove `kron_direct_sum_distr_r`.

* refactor: Get rid of `crunching...` message.

It's caused by `solve_matrix` for long messages.

* perf: Build the project in parallel.

This brings down compilation on my machine from ~75 secs -> ~45 secs.

TODO: Try to minimize the number of files that depend on one another, to
maximize parallelism. Right now the hotpath is `ExplicitDecompositions`
and `PartialTraceDefinitions`.